[FE] FIX: enum 분리 #26
