+++ src/frontend/src/components/common/button/index.css.ts
@@ -1,13 +1,5 @@
 import styled, { css } from 'styled-components';
-
-/** 버튼 색상 타입 */
-export enum ButtonColor {
-  LIGHTGRAY = '#f4f4f4',
-  GRAY = '#d4d4d4',
-  DARKGRAY = '#444444',
-  ORANGE = '#FF9100',
-  TRANSPARENT = 'transparent',
-}
+import { ButtonColor } from '@/types/enums/ButtonColor';
 
 /** 버튼 스타일 */
 export const StyledButton = styled.button<{
