+++ src/frontend/src/types/enums/ButtonColor.ts
@@ -1,0 +1,7 @@
+export enum ButtonColor {
+  LIGHTGRAY = '#f4f4f4',
+  GRAY = '#d4d4d4',
+  DARKGRAY = '#444444',
+  ORANGE = '#FF9100',
+  TRANSPARENT = 'transparent',
+}
