+++ src/frontend/src/types/enums/InputDesign.ts
@@ -1,0 +1,4 @@
+export enum InputDesign {
+  INPUT = 0,
+  SEARCH = 1,
+}
