+++ src/frontend/src/types/enums/ProfileType.ts
@@ -1,0 +1,5 @@
+export enum ProfileType {
+  INVITE = 0,
+  VOICECHAT = 1,
+  MEMBER = 2,
+}
