+++ src/frontend/src/types/enums/SidebarType.ts
@@ -1,0 +1,6 @@
+export enum SidebarType {
+  Chat = 0,
+  Playlist = 1,
+  VoiceChat = 2,
+  Member = 3,
+}
