+++ src/frontend/src/types/enums/UserRole.ts
@@ -1,0 +1,5 @@
+export enum UserRole {
+  OWNER = 0,
+  MANAGER = 1,
+  USER = 2,
+}
